refactor: AbstractPiece 클래스를 Piece 클래스로 병합
